pipeline/requirements.txt
python-telegram-bot==21.9
openai==1.68.0
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
aiohttp==3.11.12
apscheduler==3.11.0
